fastapi_app/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI Observability Demo Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-distro>=0.51b0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-instrumentation-logging>=0.51b0",
    "prometheus-client>=0.21.0",
    "uvicorn>=0.31.0",
]

